blog

+++ blog_springboot/src/main/java/com/xiaofei/blog/repository/UserDao.java
@@ -3,6 +3,16 @@ package com.xiaofei.blog.repository;
 import com.xiaofei.blog.domain.User;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface UserDao extends JpaRepository<User, Long> {
+/**
+ * @author xiaofei
+ */
+public interface UserDao extends JpaRepository<User, Integer> {
+    /**
+     * findByUsername
+     *
+     * @param username
+     * @return
+     */
     User findByUsername(String username);
+
 }
